+++ source/Domain/ValueObject/ValueMutation.ts
@@ -31,7 +31,7 @@ export class ValueMutation<T extends MutationOptions = MutationOptions> extends 
 	 */
 	get descriptor(): PropertyDescriptor {
 		const { target, key, value } = this;
-		const defaults: Partial<PropertyDescriptor> = Object.getOwnPropertyDescriptor(target, key) || { configurable: true };
+		const defaults: Partial<PropertyDescriptor> = Object.getOwnPropertyDescriptor(target, key) || { writable: true, enumerable: true, configurable: true };
 
 		return DescriptorMapper.merge(defaults, { value });
 	}